rename marktype, split data.schema.ts
