+++ tsconfig.json
@@ -40,7 +40,8 @@
         "src/enc.ts",
         "src/encdef.ts",
         "src/Encoding.ts",
-        "src/schema/marktype.ts",
+        "src/schema/data.schema.ts",
+        "src/schema/marktype.schema.ts",
         "src/schema/schema.ts",
         "src/schema/schemagen.ts",
         "src/schema/schemautil.ts",
